+++ ImageFeed/Auth/WebViewViewController.swift
@@ -36,11 +36,10 @@ final class WebViewViewController: UIViewController {
         
         estimatedProgressObservation = webView?.observe(
             \.estimatedProgress,
-             options: [],
-             changeHandler: { [weak self] _, _ in
+             options: []) { [weak self] _, _ in
                  guard let self = self else { return }
                  self.updateProgress()
-             })
+             }
     }
     
     private func updateProgress() {
