+++ ImageFeed/Models/UserResult.swift
@@ -1,0 +1,12 @@
+//
+//  UserResult.swift
+//  ImageFeed
+//
+//  Created by Dmitry Markovskiy on 28.01.2024.
+//
+
+import Foundation
+
+struct UserResult: Decodable {
+    let profileImage: ImageUrl?
+}
